moved all parsing code into pancake module
